extract setting component
